Moved location of Path class from jsonargparse to jsonargparse.typing
