+++ sphinx/index.rst
@@ -24,7 +24,7 @@ jsonargparse
 jsonargparse.typing
 -------------------
 .. automodule:: jsonargparse.typing
-    :exclude-members: get_import_path, import_object, Path
+    :exclude-members: get_import_path, import_object
 
 
 Index
